Small changes to container construction / destruction (#2)
One of the annoying things about running the integration tests is that upon stopping the tests from running, the containers are left dangling and you need to manually remove them, causing some friction in the development cycle. This PR addresses that concern by removing the container upon initialization (if the container is there obviously). You need to opt-in by setting the `AutoRemoveContainerOnInitialization` property to `true` on the docker container configuration. It's on by default for all the containers we have now.

- Code style changes (debatable but does adhere to Ironclad's style stylecop)
- AutoRemoveContainerOnInitialization added, renamed AutoRemoveContainer to AutoRemoveContainerOnDispose. Made sure auto removal upon initialization works by interrupting the tests.
- Removed the IsContainerReusable which isn't being used (we can resurface it when we have a good use case for it).